requirements.txt
flask
torch
transformers
requests
pandas
scikit-learn
python-dotenv
accelerate